First working version with mock data
This is a running version of the application. A quite large commit, but
I'll do my best to break it down into what is going on.

- Xamarin + C#
- A Portable Class Library (PCL) which holds the business logic, data
access (repository) and the Models of our data.

- A Xamarin.Android project (anmed MarketAmerica.Droid). This contains
all of the Android specific logic to build out the Activities, the
Views, and so on.

Currently, with this commit, if you install the APK you can
1. View a list of all of the Automotive sub-categories
2. Click on a Sub Category and then you will see all of the products it
contains
3. Click on an individual product to view a Detail view of the product
itself

ALL DATA IS MOCKED CURRENTLY!